.charles/pyproject.toml
[project]
name = "MLX8-W5-Audio-Transformer"
version = "0.1.0"
description = "Audio CNN and Transformer"
authors = [
    { name="Charles Cai", email="charles.cai@socialogix.net" }
]
dependencies = [
    "torch",
    "torchaudio",
    "tqdm",
    "matplotlib",
    "pandas",
    "python-dotenv",
    "numpy",
    "pyarrow",
    "soundfile>=0.13.1",
    "colorlog>=6.9.0",
    "wandb>=0.21.0",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "mido>=1.3.3",
    "pyfluidsynth>=1.3.4",
    "ipykernel>=6.29.5",
    "transformers>=4.53.1",
    "music21>=9.7.1",
]
requires-python = "==3.10.*"
